Entry #3
Figured out the old method of connecting the trees by a dime node (Dime because there are two closely related trees that must be accessible to each other (heads or tail; front or back).

Now just make a functional GUI and ask Zibad if I can meet with Tony on the weekend to demonstrate the system.